Add files via upload
This has all of the most current scripts, and the relevant data file.  Should be able to run with this.  If something is missing let me know.  (Variables to change are at the top of comborun, namely the path and then you can play with pathing iterations, total count, and number of realizations)